# subscriptions-service/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Subscriptions Aggregation Service",
    "description": "REST-сервис для агрегации данных об онлайн подписках пользователей",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/subscriptions/total": {
      "get": {
        "summary": "Подсчитать суммарную стоимость подписок",
        "description": "Возвращает суммарную стоимость всех подписок за выбранный период с возможной фильтрацией по user_id и service_name",
        "operationId": "calculateTotal",
        "tags": ["subscriptions"],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "description": "Начало периода (формат MM-YYYY)",
            "schema": {
              "type": "string",
              "example": "01-2025"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "description": "Конец периода (формат MM-YYYY)",
            "schema": {
              "type": "string",
              "example": "12-2025"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "description": "Фильтр по ID пользователя (UUID)",
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba"
            }
          },
          {
            "name": "service_name",
            "in": "query",
            "required": false,
            "description": "Фильтр по названию сервиса",
            "schema": {
              "type": "string",
              "example": "Yandex Plus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Суммарная стоимость",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TotalResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Суммарная стоимость подписок за период в рублях",
            "example": 4800
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки"
          }
        }
      }
    }
  }
}